+++ spicetify-marketplace/tools/chocolateyuninstall.ps1
@@ -1,30 +1,27 @@
 $ErrorActionPreference = 'Stop'
-$toolsDir              = Split-Path -parent $MyInvocation.MyCommand.Definition
 $customAppsDir         = "$env:userprofile\.spicetify\CustomApps"
 
-. "$toolsDir\utils.ps1"
-
+$spotifyInstalled = $true
 $spotify = Get-Process -Name 'spotify' -ErrorAction SilentlyContinue
+if(-not $spotify) {
+  $spotifyInstalled = $null -ne (Get-AppInstallLocation 'spotify')
+}
 
 Remove-Item `
   -Path "$customAppsDir\spicetify-marketplace" `
   -ErrorAction SilentlyContinue `
   -Force -Recurse
 
-Update-SpicetifyConfig `
-  -FileName "$env:userprofile\.spicetify\config-xpui.ini" `
-  -LinePart 'custom_apps' `
-  -ReplacePart 'spicetify-marketplace' `
-  -ReplaceBy ''
+spicetify config custom_apps spicetify-marketplace-
 
 Start-Process `
   -FilePath 'spicetify' `
   -ArgumentList 'apply' `
   -NoNewWindow
 
-if(-not $spotify) {
+if(-not $spotify -And $spotifyInstalled) {
   $sw = [System.Diagnostics.Stopwatch]::StartNew()
-  while($sw.ElapsedMilliseconds -lt 10000) {
+  while($sw.ElapsedMilliseconds -lt 5000) {
     $spotify = Get-Process 'spotify' -ErrorAction SilentlyContinue
     if($spotify) {
       Stop-Process $spotify -Force
